refactor: 유틸 함수 위치 이동
