+++ src/components/restaurant/util/reRenderRestaurantListContainer.js
