+++ src/components/util/TabButton.js
@@ -1,12 +1,13 @@
-import createElement from "./createElement";
+import createElement from "../../util/createElement";
 
-export default function createTabButton({className, text, isWishTab}) {
+export default function TabButton({className, text, isWishTab}) {
     const $tab = createElement({
         tag: "button",
         classNames: ["restaurant-tab", className],
     });
     $tab.textContent = text;
     $tab.dataset.wish = isWishTab;
+    
     return $tab;
 }
